Renamed group popup to represent functionality
